dev/requirements.txt
djangorestframework
djangorestframework-simplejwt
requests